backend/requirements.txt
# ── Web framework ──────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# ── Database ───────────────────────────────────────────────────
sqlalchemy==2.0.36
databases[asyncpg]==0.9.0
asyncpg==0.30.0
psycopg2-binary==2.9.10

# ── Auth & Security ────────────────────────────────────────────
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
PyJWT==2.10.1

# ── AI / ML ────────────────────────────────────────────────────
transformers==4.46.3
torch==2.5.1
sentence-transformers==3.3.1
huggingface-hub==0.26.3

# ── Vector DB ──────────────────────────────────────────────────
pinecone-client==5.0.1

# ── NLP ────────────────────────────────────────────────────────
nltk==3.9.1
textblob==0.18.0

# ── Exercise ML ────────────────────────────────────────────────
scikit-learn==1.6.0
joblib==1.4.2
pandas==2.2.3
numpy==1.26.4

# ── HTTP client (for Whereby API) ─────────────────────────────
httpx==0.28.0

# ── Config ─────────────────────────────────────────────────────
python-dotenv==1.0.1

